backend/pyproject.toml
[tool.poetry]
name = "travel-agent"
version = "0.1"
license = "Apache-2.0"
description = "Travel Agent Demo using Google ADK"
authors = ["User <user@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
absl-py = "^2.2.1"
google-adk = ">=0.4.0"
google-cloud-aiplatform = {extras = ["adk", "agent-engines"], version = "^1.88.0"}
vertexai = ">=0.4.0"
fastapi = "^0.115.12"
uvicorn = "^0.34.2"
pydantic = "^2.11.4"
pydantic-settings = "^2.9.1"
python-dotenv = "^1.0.0"
google-generativeai = ">=0.7.0"
typing-extensions = ">=4.8.0"
httpx = ">=0.25.0"
websockets = ">=12.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
